Passing the cleanup cycle time correctly and setting the access_count and timestamp on access

+++ cache_backend/CacheBackendBase.py
@@ -16,11 +16,11 @@ class CacheBackendBase(metaclass=ABCMeta):
     ttl: int = 0
     max_num_items: int = 0
     _cache_cleanup_thread: Any = None
-    _cache_cleanup_cycle_time: int = 0  # In seconds
+    _cache_cleanup_cycle_time: float = 0  # In seconds
     _cache_cleanup_handler = None
 
     def __init__(self, collection: Collection, ttl: int = 0, max_item_size: int = 1 * 10 ** 6,
-                 max_num_items: int = 1000, cache_cleanup_cycle_time: int = 1):
+                 max_num_items: int = 1000, cache_cleanup_cycle_time: float = 1):
         self.collection = collection
         self.max_item_size = max_item_size
         self.max_num_items = max_num_items
